feat(*): merge install tracing/metrics/logging commands to one command
