+++ app/kumactl/cmd/completion/testdata/bash.golden
@@ -3775,40 +3775,9 @@ _kumactl_install_gateway()
     noun_aliases=()
 }
 
-_kumactl_install_logging()
+_kumactl_install_observability()
 {
-    last_command="kumactl_install_logging"
-
-    command_aliases=()
-
-    commands=()
-
-    flags=()
-    two_word_flags=()
-    local_nonpersistent_flags=()
-    flags_with_completion=()
-    flags_completion=()
-
-    flags+=("--namespace=")
-    two_word_flags+=("--namespace")
-    local_nonpersistent_flags+=("--namespace")
-    local_nonpersistent_flags+=("--namespace=")
-    flags+=("--api-timeout=")
-    two_word_flags+=("--api-timeout")
-    flags+=("--config-file=")
-    two_word_flags+=("--config-file")
-    flags+=("--log-level=")
-    two_word_flags+=("--log-level")
-    flags+=("--no-config")
-
-    must_have_one_flag=()
-    must_have_one_noun=()
-    noun_aliases=()
-}
-
-_kumactl_install_metrics()
-{
-    last_command="kumactl_install_metrics"
+    last_command="kumactl_install_observability"
 
     command_aliases=()
 
@@ -3843,37 +3812,6 @@ _kumactl_install_metrics()
     local_nonpersistent_flags+=("--without-grafana")
     flags+=("--without-prometheus")
     local_nonpersistent_flags+=("--without-prometheus")
-    flags+=("--api-timeout=")
-    two_word_flags+=("--api-timeout")
-    flags+=("--config-file=")
-    two_word_flags+=("--config-file")
-    flags+=("--log-level=")
-    two_word_flags+=("--log-level")
-    flags+=("--no-config")
-
-    must_have_one_flag=()
-    must_have_one_noun=()
-    noun_aliases=()
-}
-
-_kumactl_install_tracing()
-{
-    last_command="kumactl_install_tracing"
-
-    command_aliases=()
-
-    commands=()
-
-    flags=()
-    two_word_flags=()
-    local_nonpersistent_flags=()
-    flags_with_completion=()
-    flags_completion=()
-
-    flags+=("--namespace=")
-    two_word_flags+=("--namespace")
-    local_nonpersistent_flags+=("--namespace")
-    local_nonpersistent_flags+=("--namespace=")
     flags+=("--api-timeout=")
     two_word_flags+=("--api-timeout")
     flags+=("--config-file=")
@@ -3983,9 +3921,7 @@ _kumactl_install()
     commands+=("crds")
     commands+=("demo")
     commands+=("gateway")
-    commands+=("logging")
-    commands+=("metrics")
-    commands+=("tracing")
+    commands+=("observability")
     commands+=("transparent-proxy")
 
     flags=()
