+++ app/kumactl/cmd/install/install.go
@@ -16,9 +16,7 @@ func NewInstallCmd(pctx *kumactl_cmd.RootContext) *cobra.Command {
 	// sub-commands
 	cmd.AddCommand(newInstallControlPlaneCmd(&pctx.InstallCpContext))
 	cmd.AddCommand(newInstallCrdsCmd(&pctx.InstallCRDContext))
-	cmd.AddCommand(newInstallMetrics(pctx))
-	cmd.AddCommand(newInstallTracing(pctx))
-	cmd.AddCommand(newInstallLogging(pctx))
+	cmd.AddCommand(newInstallObservability(pctx))
 	cmd.AddCommand(newInstallDemoCmd(&pctx.InstallDemoContext))
 	cmd.AddCommand(newInstallGatewayCmd(pctx))
 	cmd.AddCommand(newInstallTransparentProxy())
